Splitting up controller tests (POC)
